fix(interfaces): reuse interface
