+++ src/models/measurement-threshold.ts
@@ -1,26 +1,18 @@
 import Joi from 'joi';
 
-import { schema as siNumberSchema, SiNumber } from './si-number';
+import { schema as thresholdSchema, Threshold } from './threshold';
 
-const schema = Joi.object().keys({
+const schema = thresholdSchema.keys({
   hashId: Joi.string().required().example('tap192'),
   pinHashId: Joi.string().required().example('e13d57'),
-  criticallyLow: siNumberSchema.allow(null).required(),
-  low: siNumberSchema.allow(null).required(),
-  high: siNumberSchema.allow(null).required(),
-  criticallyHigh: siNumberSchema.allow(null).required(),
 })
   .description('Called an "Issue trigger" in the UI. A limit on measurements of a certain quantity. If a measurement outside these limits is registered, an issue is automatically created (if there isn\'t a relevant one open yet)')
   .tag('measurementThreshold')
   .meta({ className: 'measurementThreshold' });
 
-interface MeasurementThreshold {
+interface MeasurementThreshold extends Threshold {
   hashId: string;
   pinHashId: string;
-  criticallyLow: SiNumber | null;
-  low: SiNumber | null;
-  high: SiNumber | null;
-  criticallyHigh: SiNumber | null;
 }
 
-export { schema, MeasurementThreshold };
+export { schema, Threshold, MeasurementThreshold };
